@@ -16,15 +16,16 @@
     "@fortawesome/fontawesome-free": "^6.6.0",
     "@popperjs/core": "^2.11.8",
     "@tailwindcss/forms": "0.2.1",
-    "@vue/compiler-sfc": "3.0.7",
+    "@vue/compiler-sfc": "^3.4.38",
     "axios": "^1.7.6",
     "bootstrap": "^5.3.3",
     "chart.js": "2.9.4",
     "core-js": "3.9.1",
     "gulp": "4.0.2",
     "gulp-append-prepend": "1.0.8",
+    "pinia": "^2.2.2",
     "tailwindcss": "2.0.4",
-    "vue": "3.0.7",
+    "vue": "^3.4.38",
     "vue-router": "4.0.5"
   },
   "devDependencies": {
@@ -37,8 +38,7 @@
     "cross-env": "^7.0.3",
     "eslint": "7.22.0",
     "eslint-plugin-vue": "7.7.0",
-    "postcss": "8.4.39",
-    "vue-template-compiler": "2.6.12"
+    "postcss": "8.4.39"
   },
   "eslintConfig": {
     "root": true,
